+++ testing/rspec/spec/spec_helper.rb
@@ -3,15 +3,18 @@ unless RUBY_VERSION.to_s < '1.9.0'
   SimpleCov.command_name('cuke_cataloger-cucumber')
 end
 
+here = File.dirname(__FILE__)
+
 
 require 'cuke_cataloger'
 
+require 'tempfile'
 
 RSpec.configure do |config|
   config.before(:all) do
-    here = File.dirname(__FILE__)
     @default_file_directory = "#{here}/temp_files"
     @default_test_file_directory = "#{here}/test_files"
+    @lib_directory = "#{here}/../../../lib"
   end
 
   config.before(:each) do